typescript remove-password-from-response-data
